Address small review comments
