+++ screens/HotspotDetailScreen.tsx
@@ -26,7 +26,7 @@ import { Cat } from '../models/Cat';
 import { HotspotDetails, HotspotStatus } from '../models/Hotspot';
 import { Region, RootStackParamList } from '../types';
 import { isSmallScreen } from '../utils/helperFunc';
-import { loadHotspotDetails } from '../utils/hotspots';
+import { formatHotspotAddress, loadHotspotDetails } from '../utils/hotspots';
 
 const getStyles = (theme: ReactNativePaper.Theme) =>
   StyleSheet.create({
@@ -237,18 +237,8 @@ export const HotspotDetailScreen = ({
         </View>
       </>
     );
-
-  const hotspotDetailsValues = [
-    hotspotDetails.address,
-    hotspotDetails.city,
-    hotspotDetails.zip,
-  ];
-  const address = hotspotDetailsValues
-    .filter(x => !!x)
-    .map(x => x.trim())
-    .join(', ')
-    .trim()
-    .replace(/^, /g, '');
+ 
+  const address = formatHotspotAddress(hotspotDetails, false);
 
   return (
     <>
